requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
openai-whisper==20231117
yt-dlp==2024.08.06
python-multipart==0.0.9
